requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
duckdb==0.9.2
python-dotenv==1.0.0
pydantic==2.5.0
garminconnect==0.2.8

